Problem: Not possible to create instance of wallet client
Solution: Simplified constructors of wallet client
