+++ client-common/Cargo.toml
@@ -11,9 +11,14 @@ failure = "0.1"
 miscreant = "0.4"
 blake2 = "0.8"
 rlp = "0.3"
+hex = "0.3"
+base64 = "0.10"
 serde = { version = "1.0", features = ["derive"] }
 chrono = { version = "0.4", features = ["serde"] }
 sled = { version = "0.22", optional = true }
+jsonrpc = { version = "0.11", optional = true }
+serde_json = { version = "1.0", optional = true }
 
 [features]
-default = ["sled"]
+default = ["sled", "rpc"]
+rpc = ["jsonrpc", "serde_json"]
