--- client-core/src/tests.rs
@@ -1,5 +1,0 @@
-#![cfg(test)]
-
-mod mock_index;
-
-pub use mock_index::MockIndex;
