+++ client-index/src/index/default_index.rs
@@ -1,7 +1,3 @@
-#![cfg(all(feature = "sled", feature = "rpc"))]
-
-use std::path::Path;
-
 use chrono::offset::Utc;
 use chrono::DateTime;
 
@@ -10,53 +6,50 @@ use chain_core::tx::data::address::ExtendedAddr;
 use chain_core::tx::data::{Tx, TxId};
 use chain_core::tx::TxAux;
 use client_common::balance::{BalanceChange, TransactionChange};
-use client_common::storage::SledStorage;
-use client_common::{ErrorKind, Result};
+use client_common::tendermint::Client;
+use client_common::{ErrorKind, Result, Storage};
 
 use crate::service::*;
-use crate::tendermint::Client;
-#[cfg(not(test))]
-use crate::tendermint::RpcClient;
-#[cfg(test)]
-use crate::tests::MockClient;
 use crate::Index;
 
-/// Transaction index backed by `sled` embedded database and `RpcClient`
-pub struct RpcSledIndex {
-    address_service: AddressService<SledStorage>,
-    balance_service: BalanceService<SledStorage>,
-    global_state_service: GlobalStateService<SledStorage>,
-    transaction_outputs_service: TransactionOutputsService<SledStorage>,
-    transaction_service: TransactionService<SledStorage>,
-    #[cfg(not(test))]
-    client: RpcClient,
-    #[cfg(test)]
-    client: MockClient,
+/// Default implementation of transaction index using `Storage` and `TendermintClient`
+#[derive(Default, Clone)]
+pub struct DefaultIndex<S, C>
+where
+    S: Storage,
+    C: Client,
+{
+    address_service: AddressService<S>,
+    balance_service: BalanceService<S>,
+    global_state_service: GlobalStateService<S>,
+    transaction_outputs_service: TransactionOutputsService<S>,
+    transaction_service: TransactionService<S>,
+    client: C,
 }
 
-impl RpcSledIndex {
-    /// Creates a new instance of `RpcSledIndex`
-    pub fn new<P: AsRef<Path>>(path: P, url: &str) -> Result<Self> {
-        #[cfg(not(test))]
-        let storage = SledStorage::new(path)?;
-        #[cfg(test)]
-        let storage = SledStorage::temp(path)?;
-
-        #[cfg(not(test))]
-        let client = RpcClient::new(url);
-        #[cfg(test)]
-        let client = MockClient::new(url);
-
-        Ok(Self {
+impl<S, C> DefaultIndex<S, C>
+where
+    S: Storage + Clone,
+    C: Client,
+{
+    /// Creates a new instance of `DefaultIndex`
+    pub fn new(storage: S, client: C) -> Self {
+        Self {
             address_service: AddressService::new(storage.clone()),
             balance_service: BalanceService::new(storage.clone()),
             global_state_service: GlobalStateService::new(storage.clone()),
             transaction_outputs_service: TransactionOutputsService::new(storage.clone()),
             transaction_service: TransactionService::new(storage),
             client,
-        })
+        }
     }
+}
 
+impl<S, C> DefaultIndex<S, C>
+where
+    S: Storage,
+    C: Client,
+{
     /// Clears all the services
     fn clear(&self) -> Result<()> {
         self.address_service.clear()?;
@@ -153,7 +146,11 @@ impl RpcSledIndex {
     }
 }
 
-impl Index for RpcSledIndex {
+impl<S, C> Index for DefaultIndex<S, C>
+where
+    S: Storage,
+    C: Client,
+{
     fn sync(&self) -> Result<()> {
         let last_block_height = match self.global_state_service.last_block_height()? {
             None => {
@@ -207,11 +204,77 @@ mod tests {
 
     use std::str::FromStr;
 
+    use chrono::DateTime;
+
     use chain_core::init::address::RedeemAddress;
+    use chain_core::init::coin::Coin;
+    use chain_core::init::config::{ERC20Owner, InitConfig};
+    use client_common::storage::MemoryStorage;
+    use client_common::tendermint::types::*;
+
+    /// Mock tendermint client
+    #[derive(Default, Clone)]
+    pub struct MockClient;
+
+    impl Client for MockClient {
+        fn genesis(&self) -> Result<Genesis> {
+            Ok(Genesis {
+                genesis: GenesisInner {
+                    genesis_time: DateTime::from_str("2019-04-09T09:33:10.592188Z").unwrap(),
+                    chain_id: "test-chain-4UIy1Wab".to_owned(),
+                    app_state: InitConfig::new(vec![ERC20Owner::new(
+                        RedeemAddress::from_str("0x1fdf22497167a793ca794963ad6c95e6ffa0b971")
+                            .unwrap(),
+                        Coin::new(10000000000000000000).unwrap(),
+                    )]),
+                },
+            })
+        }
+
+        fn status(&self) -> Result<Status> {
+            Ok(Status {
+                sync_info: SyncInfo {
+                    latest_block_height: "1".to_owned(),
+                },
+            })
+        }
+
+        fn block(&self, _: u64) -> Result<Block> {
+            Ok(Block {
+            block: BlockInner {
+                header: Header {
+                    height: "1".to_owned(),
+                    time: DateTime::from_str("2019-04-09T09:38:41.735577Z").unwrap(),
+                },
+                data: Data {
+                    txs: vec!["+JWA+Erj4qBySKi4J+krjuZi++QuAnQITDv9YzjXV0RcDuk+S7pMeIDh4NaAlHkGYaL9naP+5TyquAhZ7K4SWiCliAAA6IkEI8eKw4GrwPhG+ESAAbhASZdu2rJI4Et7q93KedoEsTVFUOCPt8nyY0pGOqixhI4TvORYPVFmJiG+Lsr6L1wmwBLIwxJenWTyKZ8rKrwfkg==".to_owned()]
+                }
+            }
+        })
+        }
+
+        fn block_results(&self, _: u64) -> Result<BlockResults> {
+            Ok(BlockResults {
+                height: "2".to_owned(),
+                results: Results {
+                    deliver_tx: vec![DeliverTx {
+                        tags: vec![Tag {
+                            key: "dHhpZA==".to_owned(),
+                            value: "kOzcmhZgAAaw5roBdqDNniwRjjKNe+foJEiDAOObTDQ=".to_owned(),
+                        }],
+                    }],
+                },
+            })
+        }
+
+        fn broadcast_transaction(&self, _: &[u8]) -> Result<()> {
+            Ok(())
+        }
+    }
 
     #[test]
     fn check_flow() {
-        let index = RpcSledIndex::new("./index-test".to_owned(), "dummy").unwrap();
+        let index = DefaultIndex::new(MemoryStorage::default(), MockClient::default());
 
         let spend_address = ExtendedAddr::BasicRedeem(
             RedeemAddress::from_str("1fdf22497167a793ca794963ad6c95e6ffa0b971").unwrap(),
