+++ client-index/src/lib.rs
@@ -3,9 +3,6 @@
 
 pub mod index;
 pub mod service;
-pub mod tendermint;
-#[cfg(test)]
-pub mod tests;
 
 #[doc(inline)]
 pub use crate::index::Index;
