--- client-index/src/tests.rs
@@ -1,5 +1,0 @@
-#![cfg(test)]
-
-mod mock_client;
-
-pub use mock_client::MockClient;
